Develop

+++ Fermento/recipe_manager/views.py
@@ -12,7 +12,10 @@ from django.core import serializers
 from django.utils.dateparse import parse_duration
 import math
 from django.apps import apps
-
+from PIL import Image
+from io import BytesIO
+from django.core.files.uploadedfile import InMemoryUploadedFile
+import uuid
 
 @login_required(login_url='/accounts/login/')
 def index(request):
@@ -217,14 +220,15 @@ def edit_recipe_post(request, recipe_id):
     new_recipe.name = data["name"]
     new_recipe.description = data["description"]
     new_recipe.rating = int(data["rating"])
+    #TODO: Optionally reduce image size to reduce required space
     if "image" in request.FILES:
-        new_recipe.image = request.FILES["image"]
+        new_recipe.image = downsize_image(request.FILES["image"])
     new_recipe.difficulty = data["difficulty"]
     new_recipe.save()
     processes = json.loads(request.POST.dict()["processes"])
     db_processes = process.objects.filter(owner=o, related_recipe=new_recipe)
     delete_processes = [dbp.id for dbp in db_processes if dbp.id not in [int(p["id"]) for p in processes]]
-    process.objects.filter(id__in=delete_processes).delete()
+    process.objects.filter(id__in=delete_processes, owner=request.user).delete()
     for p in processes:
         if p["id"] == "-1":
             new_process = process()
@@ -235,25 +239,24 @@ def edit_recipe_post(request, recipe_id):
         new_process.name = p["name"]
         new_process.work_duration = parse_duration(p["work_duration"])
         new_process.wait_duration = parse_duration(p["wait_duration"])
-        print(new_process.wait_duration)
         new_process.save()
 
         #Delete ingredients that were removed by user
         db_ingredients = recipe_ingredient.objects.filter(owner=o, related_process=new_process)
         delete_ingredients = [dbi.id for dbi in db_ingredients if dbi.id not in [int(i["id"]) for i in p["ingredients"]]]
-        recipe_ingredient.objects.filter(id__in=delete_ingredients).delete()
+        recipe_ingredient.objects.filter(id__in=delete_ingredients, owner=request.user).delete()
         #Delete Utensils that were removed by user
         db_utils = utensils.objects.filter(owner=o, related_process=new_process)
         delete_utils = [dbu.id for dbu in db_utils if dbu.id not in [int(u["id"]) for u in p["utils"]]]
-        utensils.objects.filter(id__in=delete_utils).delete()
+        utensils.objects.filter(id__in=delete_utils, owner=request.user).delete()
         #Delete Process steps that were removed by user
         db_steps = process_step.objects.filter(owner=o, related_process=new_process)
         delete_steps = [dbps.id for dbps in db_steps if dbps.id not in [int(ps["id"]) for ps in p["steps"]]]
-        process_step.objects.filter(id__in=delete_steps).delete()
+        process_step.objects.filter(id__in=delete_steps, owner=request.user).delete()
         #Delete Schedules that were removed by user
         db_schedules = process_schedule.objects.filter(owner=o, related_process=new_process)
         delete_schedules = [dbs.id for dbs in db_schedules if dbs.id not in [int(s["id"]) for s in p["schedules"]]]
-        process_schedule.objects.filter(id__in=delete_schedules).delete()
+        process_schedule.objects.filter(id__in=delete_schedules, owner=request.user).delete()
 
         for i in p["ingredients"]:
             if i["id"] == "-1":
@@ -297,7 +300,7 @@ def edit_recipe_post(request, recipe_id):
             new_schedule.wait_time = timedelta(seconds=_input_to_deltatime(s["frequency"]))
             new_schedule.end_time = timedelta(seconds=_input_to_deltatime(s["end"]))
             new_schedule.save()
-    [batch.create_next_executions() for batch in apps.get_model('batches', 'Batch').objects.filter(related_recipe=new_recipe)]
+    [batch.create_next_executions() for batch in apps.get_model('batches', 'Batch').objects.filter(related_recipe=new_recipe, owner=request.user)]
     return JsonResponse({'status':'success', 'recipe_id':new_recipe.id})
 
 @login_required(login_url='/accounts/login/')
@@ -324,7 +327,7 @@ def recipe_create_post(request):
     new_recipe.description = data["description"]
     new_recipe.rating = int(data["rating"])
     if "image" in request.FILES:
-        new_recipe.image = request.FILES["image"]
+        new_recipe.image = downsize_image(request.FILES["image"])
     new_recipe.difficulty = data["difficulty"]
     new_recipe.save()
 
@@ -396,4 +399,16 @@ def _input_to_deltatime(input_string):
     total_seconds += seconds
     return total_seconds
 
-    
+def downsize_image(image):
+    if not settings.DOWNSIZE_IMAGES:
+        return image
+    basewidth = settings.MAX_IMAGE_WIDTH
+    img = Image.open(image)
+    wpercent = (basewidth/float(img.size[0]))
+    hsize = int((float(img.size[1])*float(wpercent)))
+    img = img.resize((basewidth,hsize), Image.Resampling.LANCZOS)
+    buffer = BytesIO()
+    img.save(buffer, format="JPEG")
+    buffer.seek(0)
+    image_file = InMemoryUploadedFile(buffer, None, f"image_{uuid.uuid4()}.jpg", "image/jpeg", buffer.getbuffer().nbytes, None)
+    return image_file
